chore(ipc): excise the dead broker-v2 surface (#1001)
The running-process v2 broker API (broker_v2.rs) was dead code: no
production call sites, all #[doc(hidden)] (#844), and the v2 loader
never shipped. The v2 strategy was abandoned when zccache went embedded
(#777). This removes it.

- Delete broker_v2.rs (connect_v2_broker / adopt_v2_session /
  into_backend_io_v2 + async bridges + deadline helpers) and the 6
  v2-only integration tests.
- Extract the one live piece — probe_local_socket, the raw-socket
  reachability probe used by the RUNNING_PROCESS_FAKE_BACKEND seam —
  into a focused new ipc::probe module (2 unit tests).
- Remove the v2 manifest dual-write: build_manifest_builder_v2, the v2
  writes in publish_manifest / publish_manifest_in (one fewer
  per-startup registry write), the 2 internal v2 tests, and the
  protocol_v2 import.
- Drop the now-unused interprocess dev-dep (was spike-test-only).

Net: ~800 LOC of dead code + a per-daemon-startup registry write gone,
no behavior change on the live v1 path.

Validated: zccache-ipc --all-targets green on Linux docker (42 tests);
both zccache bins compile; no dangling refs.

Closes #1001
