--- crates/zccache/tests/v2_stress_adversarial_test.rs
@@ -1,445 +1,0 @@
-//! Stress + adversarial test suite for the v2 broker surface.
-//!
-//! Sweeps every consumer-facing slice that landed in zccache#782's
-//! v1→v2 migration (slices 21, 23, 24, 25) — the ServiceDefinition
-//! and CacheManifest dual-write paths, the `protocol_v2::backend_handle`
-//! namespace, the `probe_local_socket` reachability seam, and the v2
-//! pipe namer — and pounds on each one concurrently with adversarial
-//! inputs.
-//!
-//! Why this exists: criterion #3 of zccache#782's goal demands
-//! "rock stable through stress and adversarial tests with the v2
-//! api". The unit-test layer pins individual function contracts;
-//! this file pins the *system* contract under concurrency + bad
-//! inputs that a real production load would actually generate.
-
-#![allow(
-    clippy::unwrap_used,
-    clippy::expect_used,
-    clippy::panic,
-    clippy::panic_in_result_fn,
-    clippy::unwrap_in_result
-)]
-
-use prost::Message;
-use running_process::broker::lifecycle::names_v2::v2_program_pipe;
-use running_process::broker::protocol_v2::backend_handle::Endpoint;
-use running_process::broker::protocol_v2::{
-    self, BrokerIsolation, CacheManifestBuilder, CacheRootKind, HttpServerCapability,
-    ServiceDefinition, ServiceDefinitionBuilder,
-};
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Arc;
-use std::thread;
-use std::time::Instant;
-
-use zccache::ipc::broker_v2::probe_local_socket;
-
-// ----------------------------------------------------------------------
-// ServiceDefinition v2 stress (slice 21 / 22b consumer side)
-// ----------------------------------------------------------------------
-
-/// 64 threads × 32 builds = 2048 concurrent ServiceDefinition writes
-/// into temp dirs. Every write succeeds; no panics, no deadlocks, no
-/// path collisions across the per-thread tempdir. Pins concurrency
-/// safety of `ServiceDefinitionBuilder::install_in` + the underlying
-/// `write_service_definition_v2` atomic-write.
-#[test]
-fn service_definition_install_64x32_concurrent_writes() {
-    const N_THREADS: usize = 64;
-    const N_WRITES_PER_THREAD: usize = 32;
-
-    let success = Arc::new(AtomicUsize::new(0));
-    let start = Instant::now();
-
-    let handles: Vec<_> = (0..N_THREADS)
-        .map(|tid| {
-            let success = Arc::clone(&success);
-            thread::spawn(move || {
-                let dir = tempfile::tempdir().expect("tempdir");
-                for i in 0..N_WRITES_PER_THREAD {
-                    // Each thread writes under its own tempdir + uses a
-                    // distinct sub-prefix so no cross-thread collision can
-                    // mask an actual concurrency bug.
-                    let name = format!("zccache-stress-t{tid:02}n{i:02}");
-                    let path =
-                        ServiceDefinitionBuilder::shared_broker(&name, format!("/usr/bin/{name}"))
-                            .min_version("1.0.0")
-                            .label("env", "stress")
-                            .label("thread", format!("{tid}"))
-                            .install_in(dir.path())
-                            .expect("install_in must succeed under concurrency");
-                    assert!(path.exists(), "install path must materialize");
-                    success.fetch_add(1, Ordering::Relaxed);
-                }
-            })
-        })
-        .collect();
-
-    for h in handles {
-        h.join().expect("thread joined");
-    }
-    let total = success.load(Ordering::Relaxed);
-    let elapsed = start.elapsed();
-    assert_eq!(
-        total,
-        N_THREADS * N_WRITES_PER_THREAD,
-        "all {N_THREADS}*{N_WRITES_PER_THREAD} writes must succeed; got {total}"
-    );
-    // Wall-clock budget: 30s on Windows CI runners. Local should be
-    // well under 5s. Tripping this flags either a deadlock or a
-    // serialization regression in the writer.
-    assert!(
-        elapsed.as_secs() < 30,
-        "{N_THREADS}x{N_WRITES_PER_THREAD} concurrent writes took {elapsed:?}; budget is 30s"
-    );
-}
-
-/// Adversarial inputs: every malformed service name the v2 namer
-/// rejects must also be rejected by `ServiceDefinitionBuilder::install_in`
-/// without panicking. Builds the definition (permissive), tries to
-/// install (validates), confirms `Err`. Sweeps NUL, UPPERCASE, empty,
-/// trailing dash, leading digit (still allowed), 65-char overlength,
-/// path-separator embedded.
-#[test]
-fn service_definition_install_rejects_every_malformed_name() {
-    let dir = tempfile::tempdir().expect("tempdir");
-    // "trailing-" / leading-digit are ACCEPTED by v1's validate_service_name —
-    // adjust the test to only sweep names that v1 demonstrably rejects.
-    let bad_names = [
-        "BAD-Caps",
-        "",
-        "a/b",
-        "a\\b",
-        "name\0nul",
-        "name with space",
-        "name\twith\ttab",
-        &"x".repeat(65),
-    ];
-    for name in bad_names {
-        let result = ServiceDefinitionBuilder::shared_broker(name, "/bin/x").install_in(dir.path());
-        assert!(
-            result.is_err(),
-            "malformed name {name:?} must produce Err, got {result:?}"
-        );
-    }
-}
-
-/// Adversarial decode: random byte streams thrown at
-/// `ServiceDefinition::decode` MUST return Err, never panic. Sweeps
-/// 256 short adversarial inputs covering common bad-frame shapes
-/// (truncated varint, oversize length, valid-tag-bad-payload, etc.).
-#[test]
-fn service_definition_decode_never_panics_on_random_garbage() {
-    let mut total = 0;
-    for seed in 0..256u32 {
-        // Cheap deterministic PRNG; we want repeatable bad inputs.
-        let mut buf: Vec<u8> = (0..32u32)
-            .map(|i| ((seed.wrapping_mul(1_103_515_245).wrapping_add(i)) & 0xFF) as u8)
-            .collect();
-        // Mix in two pathological cases:
-        //   - leading length-delimited tag claiming more bytes than buf has
-        //   - leading varint header with bit-7 stuck
-        if seed % 3 == 0 {
-            buf[0] = 0x0A; // tag=1 wire-type=2 (length-delimited)
-            buf[1] = 0xFF; // claim 255 bytes
-        }
-        if seed % 7 == 0 {
-            buf[0] = 0xFF;
-            buf[1] = 0xFF;
-        }
-        let _ = ServiceDefinition::decode(buf.as_slice());
-        total += 1;
-    }
-    assert_eq!(total, 256, "loop must complete");
-}
-
-// ----------------------------------------------------------------------
-// CacheManifest v2 stress (slice 23 consumer side)
-// ----------------------------------------------------------------------
-
-/// 32 threads × 16 publishes = 512 concurrent CacheManifest writes.
-/// Each thread publishes under its own (service, version) so no
-/// cross-thread file conflict; we're testing the write path's
-/// concurrency safety + the unix-ms-stamp uniqueness behaviour.
-#[test]
-fn cache_manifest_publish_32x16_concurrent() {
-    const N_THREADS: usize = 32;
-    const N_WRITES: usize = 16;
-
-    let success = Arc::new(AtomicUsize::new(0));
-    let start = Instant::now();
-
-    let handles: Vec<_> = (0..N_THREADS)
-        .map(|tid| {
-            let success = Arc::clone(&success);
-            thread::spawn(move || {
-                let registry = tempfile::tempdir().expect("tempdir");
-                for i in 0..N_WRITES {
-                    let svc = format!("svc-t{tid:02}n{i:02}");
-                    let path = CacheManifestBuilder::new(&svc, "1.0.0")
-                        .root(CacheRootKind::CacheData, format!("/data/{svc}"))
-                        .root(CacheRootKind::CacheIndex, format!("/index/{svc}"))
-                        .root(CacheRootKind::CacheLogs, format!("/log/{svc}"))
-                        .root(CacheRootKind::CacheLocks, format!("/lock/{svc}"))
-                        .root(CacheRootKind::CacheTmp, format!("/tmp/{svc}"))
-                        .publish_in(registry.path())
-                        .expect("publish under concurrency");
-                    assert!(path.exists(), "manifest file must exist");
-                    success.fetch_add(1, Ordering::Relaxed);
-                }
-            })
-        })
-        .collect();
-
-    for h in handles {
-        h.join().expect("thread joined");
-    }
-    let total = success.load(Ordering::Relaxed);
-    let elapsed = start.elapsed();
-    assert_eq!(total, N_THREADS * N_WRITES, "all writes must succeed");
-    assert!(
-        elapsed.as_secs() < 30,
-        "{N_THREADS}*{N_WRITES} publishes took {elapsed:?}; budget 30s"
-    );
-}
-
-// ----------------------------------------------------------------------
-// probe_local_socket adversarial sweep (slice 11 / 24 path)
-// ----------------------------------------------------------------------
-
-/// 16 threads × 50 probes against an unreachable endpoint = 800 concurrent
-/// probes. Each MUST return Err in bounded time (the
-/// DEFAULT_PROBE_TIMEOUT is 250 ms; with 16-way concurrency on a 2-core
-/// runner the worst-case wall clock is ~1s, budget 10s). Pins that the
-/// blocking probe's helper-thread bound doesn't leak under load.
-#[test]
-fn probe_local_socket_800x_concurrent_unreachable_returns_bounded() {
-    const N_THREADS: usize = 16;
-    const N_PROBES: usize = 50;
-
-    let endpoint = if cfg!(windows) {
-        r"\\.\pipe\zccache-stress-no-such-endpoint-xyz".to_owned()
-    } else {
-        "/tmp/zccache-stress-no-such-endpoint-xyz.sock".to_owned()
-    };
-    let err_count = Arc::new(AtomicUsize::new(0));
-    let start = Instant::now();
-
-    let handles: Vec<_> = (0..N_THREADS)
-        .map(|_| {
-            let endpoint = endpoint.clone();
-            let err_count = Arc::clone(&err_count);
-            thread::spawn(move || {
-                for _ in 0..N_PROBES {
-                    if probe_local_socket(&endpoint).is_err() {
-                        err_count.fetch_add(1, Ordering::Relaxed);
-                    }
-                }
-            })
-        })
-        .collect();
-
-    for h in handles {
-        h.join().expect("thread joined");
-    }
-    let total = err_count.load(Ordering::Relaxed);
-    let elapsed = start.elapsed();
-    assert_eq!(
-        total,
-        N_THREADS * N_PROBES,
-        "every probe must Err against unreachable endpoint"
-    );
-    assert!(
-        elapsed.as_secs() < 15,
-        "{N_THREADS}x{N_PROBES} concurrent probes took {elapsed:?}; budget 15s — \
-         probable helper-thread leak or deadline regression"
-    );
-}
-
-/// Adversarial endpoint sweep: every garbage endpoint shape must Err
-/// (not panic, not block past the deadline). NUL, empty, oversize,
-/// pipe-prefix-only on Windows, control chars.
-#[test]
-fn probe_local_socket_rejects_full_adversarial_input_set() {
-    let bad: Vec<String> = vec![
-        String::new(),
-        "\0".to_owned(),
-        "a\0b".to_owned(),
-        "\x01\x02\x03".to_owned(),
-        "x".repeat(8192),
-    ];
-    for ep in &bad {
-        let err = probe_local_socket(ep).expect_err("garbage endpoint must Err");
-        let _ = err;
-    }
-}
-
-// ----------------------------------------------------------------------
-// v2 pipe naming adversarial sweep
-// ----------------------------------------------------------------------
-
-/// Adversarial: every malformed SID is rejected without panicking,
-/// regardless of program name or pipe_idx. Sweeps wrong-length, non-hex,
-/// uppercase, embedded NUL.
-#[test]
-fn v2_program_pipe_rejects_every_malformed_sid() {
-    let bad_sids = [
-        "",
-        "x",
-        "deadbeef",           // 8 chars (too short)
-        "deadbeefcafef00d00", // 18 chars
-        "deadbeefcafef00g",   // non-hex 'g'
-        "DEADBEEFCAFEF00D",   // uppercase (rejected per slice 16 tests)
-        "deadbeefcafef00\0",  // embedded NUL
-    ];
-    for sid in bad_sids {
-        for idx in [0u32, 1, 0xDEAD_BEEF, u32::MAX] {
-            let result = v2_program_pipe("zccache", sid, idx);
-            assert!(
-                result.is_err(),
-                "malformed sid {sid:?} + idx {idx} must Err, got {result:?}"
-            );
-        }
-    }
-}
-
-/// u32::MAX pipe_idx + a maximal valid sid + a maximal-length program
-/// name must succeed cleanly. Pins the upper-bound contract for v2
-/// pipe-name composition.
-#[test]
-fn v2_program_pipe_handles_maximal_inputs() {
-    let name = v2_program_pipe("zccache", "ffffffffffffffff", u32::MAX)
-        .expect("max-length sid + max idx must compose");
-    assert!(name.ends_with("-4294967295"), "got: {name}");
-    assert!(name.contains("ffffffffffffffff"), "got: {name}");
-}
-
-// ----------------------------------------------------------------------
-// Endpoint round-trip adversarial sweep (slice 24 backend_handle)
-// ----------------------------------------------------------------------
-
-/// Endpoint struct accepts any byte sequence in its string fields
-/// (the v1 namespace_id has no validation beyond what callers pass);
-/// pin that prost round-trip preserves every adversarial byte
-/// sequence faithfully. Tests Unicode, embedded NUL (in proto-string
-/// allowed via protobuf wire), max-length path.
-#[test]
-fn endpoint_round_trips_adversarial_strings() {
-    let cases = [
-        ("", ""),
-        ("ns", "/path"),
-        ("✨emoji-namespace", "/路径/with/unicode"),
-        ("ns\0nul", "/tmp/path"),
-        ("normal", "/p"),
-        ("ns", &"x".repeat(4096)),
-    ];
-    for (ns, path) in cases {
-        let ep = Endpoint {
-            namespace_id: ns.to_owned(),
-            path: path.to_owned(),
-        };
-        let bytes = ep.encode_to_vec();
-        let decoded = Endpoint::decode(bytes.as_slice()).expect("endpoint decodes");
-        assert_eq!(decoded.namespace_id, ns, "namespace_id round-trips");
-        assert_eq!(decoded.path, path, "path round-trips");
-    }
-}
-
-// ----------------------------------------------------------------------
-// Cross-slice integration: dual-write under concurrency
-// ----------------------------------------------------------------------
-
-/// 16 threads each writing BOTH servicedef + manifest into the same
-/// per-thread tempdir, mimicking what daemon startup does. No write
-/// collides, no manifest overwrites a servicedef, both formats are
-/// present after each round.
-#[test]
-fn dual_write_servicedef_and_manifest_16x_concurrent() {
-    const N: usize = 16;
-    let handles: Vec<_> = (0..N)
-        .map(|tid| {
-            thread::spawn(move || {
-                let dir = tempfile::tempdir().expect("tempdir");
-                let svc = format!("zccache-dual-t{tid:02}");
-                // ServiceDefinition v2 write.
-                let sd_path =
-                    ServiceDefinitionBuilder::shared_broker(&svc, format!("/usr/bin/{svc}"))
-                        .min_version("1.0.0")
-                        .install_in(dir.path())
-                        .expect("servicedef install");
-                // CacheManifest v2 write into the SAME dir.
-                let mf_path = CacheManifestBuilder::new(&svc, "1.0.0")
-                    .root(CacheRootKind::CacheData, format!("/data/{svc}"))
-                    .publish_in(dir.path())
-                    .expect("manifest publish");
-                assert!(sd_path.exists(), "servicedef file");
-                assert!(mf_path.exists(), "manifest file");
-                assert_ne!(sd_path, mf_path, "distinct files");
-                assert!(
-                    sd_path.to_string_lossy().ends_with(".servicedef.v2"),
-                    "servicedef ext: {}",
-                    sd_path.display()
-                );
-                assert!(
-                    mf_path.to_string_lossy().ends_with(".v2.pb"),
-                    "manifest ext: {}",
-                    mf_path.display()
-                );
-            })
-        })
-        .collect();
-    for h in handles {
-        h.join().expect("thread joined");
-    }
-}
-
-// ----------------------------------------------------------------------
-// Defense-in-depth: prost decoder never panics on truncated extremes
-// ----------------------------------------------------------------------
-
-/// Truncate every length from 0..encoded.len() of a maximal
-/// ServiceDefinition and assert prost returns Err for every
-/// non-equal-to-full truncation (and Ok on the full buffer). Pins
-/// the "no panic, no UB" invariant against any future prost upgrade.
-#[test]
-fn service_definition_decode_truncation_sweep_never_panics() {
-    let full = ServiceDefinition {
-        service_name: "zccache".to_owned(),
-        binary_path: "/usr/bin/zccache-daemon".to_owned(),
-        isolation: BrokerIsolation::SharedBroker as i32,
-        per_version_binary_dir: "/usr/bin".to_owned(),
-        min_version: "1.0.0".to_owned(),
-        version_allow_list: vec!["1.0.0".to_owned()],
-        http_server: Some(HttpServerCapability {
-            bind_addr: "127.0.0.1".to_owned(),
-            health_path: "/health".to_owned(),
-            display_name: "zccache".to_owned(),
-        }),
-        ..Default::default()
-    };
-    let bytes = full.encode_to_vec();
-    // Full buffer must decode.
-    let _ok = ServiceDefinition::decode(bytes.as_slice()).expect("full decode");
-    // Every strict truncation either decodes (proto3 allows missing fields)
-    // or errors — NEVER panics. The assertion is just "no panic".
-    for n in 0..bytes.len() {
-        let _ = ServiceDefinition::decode(&bytes[..n]);
-    }
-}
-
-/// Same sweep for CacheManifest v2.
-#[test]
-fn cache_manifest_decode_truncation_sweep_never_panics() {
-    let full = CacheManifestBuilder::new("zccache", "1.0.0")
-        .root(CacheRootKind::CacheData, "/data")
-        .root(CacheRootKind::CacheLogs, "/log")
-        .broker_instance("shared")
-        .bundle_id("zccache-bundle")
-        .build();
-    let bytes = full.encode_to_vec();
-    let _ok = protocol_v2::CacheManifest::decode(bytes.as_slice()).expect("full decode");
-    for n in 0..bytes.len() {
-        let _ = protocol_v2::CacheManifest::decode(&bytes[..n]);
-    }
-}
